third_party/blink/web_tests/platform/generic/accessibility/canvas-description-and-role-expected.txt

This test makes sure that a canvas with and without fallback content each has the right role and description.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axContainer.childrenCount is 2
Canvas 1 description: Canvas label
Canvas 1 role: AXRole: AXCanvasWithFallbackContent
Canvas 2 description: 
Canvas 2 role: AXRole: AXCanvasWithFallbackContent
PASS successfullyParsed is true

TEST COMPLETE
